chore: extract simpleBytes to testutil package
